fix: replace custom stringify with json-stable-stringify
