+++ package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "@lit-protocol/constants": "^7.2.0",
     "@safe-global/api-kit": "^4.0.0",
-    "@safe-global/protocol-kit": "^6.1.0"
+    "@safe-global/protocol-kit": "^6.1.0",
+    "json-stable-stringify": "^1.3.0"
   }
 }
